[refactor] avoid Redundant operations: (#107)
* Remove redundant 'isInstance()' or 'cast()' call
* Remove redundant array creation
* Remove redundant String operation
* Remove redundant type cast

+++ core/misc/image-debayer/src/main/java/esa/opssat/camera/processing/OPSSATCameraDebayering.java
@@ -67,7 +67,7 @@ public class OPSSATCameraDebayering {
         fi.width = 2048;
         fi.height = 1944;
 
-        fi.offset = (int) 0;
+        fi.offset = 0;
         fi.nImages = 1;
         fi.gapBetweenImages = 0;
         fi.intelByteOrder = true; // Little-endian byte order
